requirements.txt
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.1.1
APScheduler==3.10.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.1.0
black==23.9.1
boto3==1.34.29
botocore==1.34.29
<<<<<<< HEAD
cachetools==5.3.3
=======
>>>>>>> bff8062244edea18a336f80a7184136605502a87
celery==5.3.4
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
<<<<<<< HEAD
chardet==5.2.0
=======
>>>>>>> bff8062244edea18a336f80a7184136605502a87
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==15.1.0
<<<<<<< HEAD
crispy-bootstrap5==2024.2
=======
>>>>>>> bff8062244edea18a336f80a7184136605502a87
cron-descriptor==1.4.0
cryptography==41.0.5
daphne==4.0.0
dataclasses-json==0.5.8
Django==4.2.5
<<<<<<< HEAD
django-background-tasks==1.2.5
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-compat==1.0.15
django-cors-headers==4.2.0
django-crispy-forms==2.1
=======
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==4.2.0
>>>>>>> bff8062244edea18a336f80a7184136605502a87
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-redis==5.4.0
django-timezone-field==6.1.0
djangochannelsrestframework==1.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
et-xmlfile==1.1.0
exceptiongroup==1.2.0
Faker==12.0.1
flower==2.0.1
frozenlist==1.4.0
<<<<<<< HEAD
google-api-core==2.17.1
google-api-python-client==2.121.0
google-auth==2.28.2
google-auth-httplib2==0.2.0
googleapis-common-protos==1.62.0
gunicorn==21.2.0
h11==0.14.0
httplib2==0.22.0
=======
gunicorn==21.2.0
h11==0.14.0
>>>>>>> bff8062244edea18a336f80a7184136605502a87
humanize==4.9.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
iniconfig==2.0.0
isort==5.12.0
jmespath==1.0.1
kombu==5.3.2
marshmallow==3.20.1
marshmallow-enum==1.5.1
mixer==7.2.2
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.0
openpyxl==3.1.2
packaging==23.2
pandas==2.1.1
pathspec==0.11.2
phonepe-sdk==1.1.0
Pillow==10.0.1
platformdirs==3.11.0
pluggy==1.3.0
prometheus-client==0.19.0
prompt-toolkit==3.0.39
<<<<<<< HEAD
protobuf==4.25.3
=======
>>>>>>> bff8062244edea18a336f80a7184136605502a87
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.18.0
PyJWT==2.8.0
pyOpenSSL==23.3.0
<<<<<<< HEAD
pyparsing==3.1.2
=======
>>>>>>> bff8062244edea18a336f80a7184136605502a87
pytest==7.4.4
python-crontab==3.0.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
razorpay==1.4.1
redis==5.0.1
requests==2.31.0
<<<<<<< HEAD
rsa==4.9
s3transfer==0.10.0
service-identity==23.1.0
shortuuid==1.0.11
=======
s3transfer==0.10.0
service-identity==23.1.0
>>>>>>> bff8062244edea18a336f80a7184136605502a87
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
tornado==6.4
twilio==8.9.0
Twisted==23.10.0
txaio==23.1.1
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
<<<<<<< HEAD
uritemplate==4.1.1
urllib3==1.26.12
=======
urllib3==2.0.6
>>>>>>> bff8062244edea18a336f80a7184136605502a87
vine==5.0.0
wcwidth==0.2.8
yarl==1.9.2
zope.interface==6.1
